requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.16
django-celery-results==2.4.0
# django-debug-toolbar==3.7.0
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
importlib-metadata==5.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.3.0
prompt-toolkit==3.0.32
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.5
qiniu==7.9.0
redis==4.3.4
requests==2.28.1
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlwt==1.3.0
zipp==3.10.0
uwsgi